Prevent contentEditable element from losing focus
If there is a contentEditable DIV and a hansontable on the same page, I can't copy-paste anything into the DIV because handsontable removes the focus from the DIV when pressing the ctrl-key. See the following fiddle for a demonstration: http://jsfiddle.net/y5969sbk/

This commit adds a check for the contentEditable flag in order to prevent further processing of the keydown event.