Check for autofill existence
Depending on if the autofill plugin is turned on, an autofill object may
or may not be attached to the handsontable instance. If it is disabled,
the plugin does not attach itself, but later code assumes that is does.
This adjust the checks so that we first see if the autofill plugin is
attached to the handsontable instance.