fix instance.dataType checker inside loadData()
Inside loadData, the "instanceof" checker was being used to set the instance.dataType attribute. Unfortunately, this breaks when a pseudoclassical constructor builds an instantiation settings object with a data property that is an array. Instanceof will instead register it as an object, breaking the table in some instances. 

Array.isArray is instead better able to accurately note the type of the data property. I've used the provided polyfill/shim helper Handsontable.helper.isArray already found in the code base. 

The bug that I'm referring to can be reproduced in a jsfiddle (the alert will note the data type instance of array incorrectly being false, even though it is definitely an array). The exact point of break I was finding in my own project was in src/core 1742, which did not like that my data array was being noted as instance.dataType === 'obj'. The jsfiddle table doesnt break, but the bug fix relating to instanceof being a poor type checker, is shown in practice. http://jsfiddle.net/17Lgkoap/

This fix will avoid future problems for anyone using constructors.