Fix placement of context menu when placed "above".
Resolution to #1993, citing that when the context menu goes 'above' the
cursor/row it goes too high.  The issue seemed to be that the context
menu item wasn't being resized as needed.  Seemed to be targeting the
right item, but it was not working.  By targeting the parentElement it
fixes the issue.