Fix event handler leak when showing comment popups
The mouseover event handler is not properly released when an existing comment is shown (line 71 detects an existing comment). Another mouseover handler gets added later on inside showComment (which calls placeCommentBox, which calls bindMouseEvent, which adds the additional handler). This leads to multiple event handlers adding up to a large chain and slowing down the browser* until anoother cell is clicked (and the handlers are reset). This commit adds the missing handler release.

* IE11 even crashes after a couple of fast mouse moves between cells with existing comments, see demo page


this is inside the mouseover event handler, make sure to get rid of it, too or else it gets rebound twice when showing the comment