Allow 0 as a valid row height
I'm currently trying to work on a plugin to allow hiding rows. But simply hiding them by applying css alone isn't enough, I also have to convince the virtual native scrollbars that they are not present, by using the "modifyRowHeight" hook. Unfortunately, the way this function is written now, if the the rowheight returned is zero, it will `||` it to `defaultRowHeight` as `0` in js is falsey. Changing `defaultRowHeight` to zero causes plenty of other issue to break so it is best to fix it so that the scrollbar height correctly calculates.