Allow '.' in property names
Simple change in program flow to allow '.' (period) in a property name.

This way, if a data set has a property of "Misc. Item", it'll find and get/set it.  Currently, it'll assume all '.' mean nested properties and try to find data['Misc'][' Item'] which won't exist.

When dot notation is being used as desired, it'll fall through the hasOwnProperty condition and work normally.