Change argument types for selection hooks
After thinking about what necessary values are important to have in
selection hooks I came to conclusion to revert arguments structure
passed to the hooks. Now, the arguments types are the same as before the
breaking change. As a compatible change, this commit adds new argument
which is 'selectionLayerLevel' which indicates what selection layer is
currently modified.

Issue: #4708