Make the `alter` method more stable when indexes group are used
To make the `alter` method less buggy the normalization algorithm is
added. If the user will try to remove columns or rows using unordered
groups of indexes such as `[[4, 2], [11, 1], [5, 5], [1, 3]]` ([[index,
amount]]) the normalization process convert it to `[[1, 9], [11, 1]]`
and pass to the removing action. The normalization process takes care to
sort the passed groups and join them into bigger groups if necessary.

Issue: #4815