Add missing API for non-contiguous selection
The new method is added to public API - `selectCells`. This method
accepts coordinates passed as an array of arrays (the same format as
`getSelected()` returns) or as an array of CellRanges (the same format
as `getSelectedRange()` returns). The method allows making the
non-contiguous selection.

Add depreciation warning to `selectCellByProp` method. Now, the both
`selectCell` and `selectCells` methods support coordinates passed as
numbers or as column properties (strings).

To make selection functionality more encapsulated the two more methods
is added - `selectRows` and `selectColumns`. The places where the
internal methods of the Selection class were used replaced with the
new ones.

Issue: #4811