Fix afterSelectionEnd not being fired correctly on mobile (#4979)
* Fix afterSelectionEnd not being fired correctly on mobile

This commit fixes the issue where using a hook on afterSelectionEnd didn't work on mobile. To do that, I added an event listener to the `touchend` event, and make it end a selection if there's one. To make sure it is fired after the selection occurs, it's placed in a setTimeout of length 0. Otherwise it would sometimes fire before the selection started.

* Add test for EventManager - touchend

* Apply reviewer's comments