Extend SampleGenerator class
Extends the SampleGenerator class for `bundleCountSeed` feature which
could be used to control the cell data uniqueness. This feature is
implemented in AutoRowSize and AutoColumnSize plugin which incorrectly
calculates the width of the checkbox cell types with enabled labels.
Now, the label values take part in collecting cell data samples by
modifying the sample bundles.

Issue: #3437