Fix scroll issue
Fixes an issue causing Handson Table to scroll to the top of the page.
The issue seems to be caused by the paste event being fired before the
copy/paste plugin is enabled. When this happens, the text editor
placeholder element receives focus. Since the text editor placeholder
element was being set to -9999 top and left, this was causing the screen
to scroll to the top. Not setting a top and left value fixes the issue.