Prevent getProperty from iterating into null
When reading a column description of the form "object.inner.value", with data of form e.g. 
[ { inner:{ value:2 } }, { inner: null } ]
getProperty would return an error 'Cannot read property 'value' of null'.

Expanded the existing "=== undefined" (void 0) check to also include "=== null", as this is similarly defensive and should not be trigger an exception.