Prevent DataMap.get from iterating into null values.
When reading a column description of the form "object.inner.value", with data of form e.g.
[ { inner:{ value:2 } }, { inner: null } ]
DataMap.get would return an error 'Cannot read property 'value' of null'.

Replacing the existing "typeof out === undefined" with a "=== null" check (perversely typeof(null) == 'object'), as this is similarly defensive and should not be trigger an exception.
Changed from a typeof to a === as (imho) typeof is only really useful when dealing with variables that might be unscoped, and 'out' is a local variable declared in the same scope.
In Chrome, the === undefined is _very marginally_ faster than typeof (0.4s faster across 10^12 iterations).