Fix bug while writing the tests
Add tests for missing modules `OrderView` and `SharedOrderView`. While
writing the tests one bug was found which occurs in shared view mode.
The bug which generates the wrong number DOM elements was fixed and
covered with the tests. Additionally, an oversized row heights cache-ing
mechanism was removed from Walkontable - it is already cleared before
table rendering.