fix running "test:e2e.watch" on Git Bash for Windows
by changing single quotes to escaped double quotes, because otherwise I got the following error:

'np' is not recognized as an internal or external command,
operable program or batch file.
'run' is not recognized as an internal or external command,
operable program or batch file.
The filename, directory name, or volume label syntax is incorrect.
'--watch'' is not recognized as an internal or external command,
operable program or batch file.
'np' is not recognized as an internal or external command,
operable program or batch file.
'run' is not recognized as an internal or external command,
operable program or batch file.
The filename, directory name, or volume label syntax is incorrect.
'--watch'' is not recognized as an internal or external command,
operable program or batch file.