refactor: use instance.getCell to get a TD
because that is the most DRY way to get the cell for a relevant overlay, as suggested in https://github.com/handsontable/handsontable/pull/6200#discussion_r319089333.

This change breaks backwards compat by returning "null" instead of "undefined" when the TD is not found. I find this safe to do because, as is no test for it and no documentation anyway.