WOT Table return correct information in the overlay (#6200)
* fix "getLastRenderedRow"  to check the relevant method

it seems a bug in the tests that these two tests were actually checking "getFirstRenderedRow"

it seems strange that one assertion scrolls to make row 10 visible, and then checks that last rendered row is 11, but this is the current behavior so be it

* change "getLastRenderedRow", "isRowAfterRenderedRows" to respond in context of the overlay #6191

* amend last: test all overlays

* change "getLastVisibleRow", to respond in context of the overlay #6191

* refactor: use polymorphism to adjust Table behavior for overlays

* amend last: change two methods to be no-ops in overlays

* refactor: split "getRenderedRowsCount" into overlay table variants and add test

* refactor: split "getRenderedColumnsCount" into overlay table variants and add test

* refactor: unify ways of checking the type of overlay

* refactor: simplify "isOverlay" logic

* change "getVisibleRowsCount", to respond in context of the overlay #6191

* change "getVisibleColumnsCount", to respond in context of the overlay #6191

* bugfix: "getLastReneredRow" should take into account that the number of actual total rows is smaller than "fixedRowsTop" setting

* refactor: consistently use "this.wot" over "this.instance"

* change "getLastRenderedColumn", to respond in context of the overlay #6191

* change "getLastVisibleColumn", to respond in context of the overlay #6191

* change "getFirstRenderedRow", to respond in context of the overlay #6191

* change "getFirstVisibleRow", to respond in context of the overlay #6191

* change "getFirstRenderedColumn", to respond in context of the overlay #6191

* change "getFirstVisibleColumn", to respond in context of the overlay #6191

* change "isBeforeRenderedRows", to respond in context of the overlay #6191

* add tests for "isRowAfterViewport"

* add tests for "isRowAfterRenderedRows"

* bugfix: return value was incorrectly offsetted

* safeguard: add assertion that TR and TD was found

* refactor: use existing helper function to determine the starting row and column

* refactor: never use Table class directly

in case of "not overlay", instantiate MasterTable

* refactor: move code used by one or more Table classes to a mixin

* apend refactor: move code used by one or more Table classes to a mixin (columns)

* code style: white space and typo improvements

as suggested and explained in:
- https://github.com/handsontable/handsontable/pull/6200#discussion_r313302196
- https://github.com/handsontable/handsontable/pull/6200#discussion_r313308930

* refactor: move two methods to MasterTable

move methods `alignOverlaysWithTrimmingContainer`, `markOversizedColumnHeaders` to MasterTable,
because they were no-ops in other (overlay) tables

* fix implementation and tests of getCell and related functions

fix implementation and tests of getCell the methods:
- getCell
- isRowBeforeRenderedRows
- isRowAfterRenderedRows
- isColumnBeforeRenderedColumns
- isColumnAfterRenderedColumns

to match the expectation that this methods should work on row & column headers when given a negative index. This is to address the comment https://github.com/handsontable/handsontable/pull/6200#pullrequestreview-274163279

* make sure helper functions return -1 in case there is no rows/columns in any given overlay

affects functions:
- getFirstRenderedRow()
- getFirstVisibleRow()
- getLastRenderedRow()
- getLastVisibleRow()

to address comment https://github.com/handsontable/handsontable/pull/6200#pullrequestreview-274163279

* fix failing tests

fix few places in code that were expecting null value instead of -1 returned from getFirstRenderedRow, getFirstVisibleRow, getFirstRenderedColumn, getFirstVisibleColumn

* remove redundant class

remove redundant subclass/superclass OverlayTable, which was common for all subclasses of Table which are not MasterTable. This class not used anywhere. Removing because it is YAGNI

* add missing JSDoc

* Coding style improvements

as suggested by @swistach in the review https://github.com/handsontable/handsontable/pull/6200#pullrequestreview-281327596

Co-Authored-By: Piotr Laszczkowski <swistach@users.noreply.github.com>

* Coding style improvements #2

as suggested by @swistach in the review https://github.com/handsontable/handsontable/pull/6200#pullrequestreview-281327596

Co-Authored-By: Piotr Laszczkowski <swistach@users.noreply.github.com>

* Coding style improvements #3

as suggested by @swistach in the review https://github.com/handsontable/handsontable/pull/6200#pullrequestreview-281327596

* use async/await sleep instead of setTimeout

as suggested by @swistach in the review https://github.com/handsontable/handsontable/pull/6200#pullrequestreview-281327596

* refactor: use instance.getCell to get a TD

because that is the most DRY way to get the cell for a relevant overlay, as suggested in https://github.com/handsontable/handsontable/pull/6200#discussion_r319089333.

This change breaks backwards compat by returning "null" instead of "undefined" when the TD is not found. I find this safe to do because, as is no test for it and no documentation anyway.

* improve JSDoc

because it was missing information about the order of validation of the input parameters, as pointed out in https://github.com/handsontable/handsontable/pull/6200#discussion_r319079958

* code cleanup: put helper methods on top of the file

because they are or might be used multiple times

* add missing test for using hot.getCell with Nested Headers

this test checks that the first nested header row has index -1, the next one is -2, etc. The discussion about the desired order is here: https://github.com/handsontable/handsontable/pull/6200#discussion_r319050906

* change the implementation of negative row index in getCell

now, with nested headers, "-1" means the inner header and "-<countHeaders>" means the outmost header

as requested in https://github.com/handsontable/handsontable/pull/6200#discussion_r319050906

* refactor: simplify and document how the function behaves for negative indexes

as pointed out in https://github.com/handsontable/handsontable/pull/6200#discussion_r319064155

* bugfix: the role of negative indexes was changed

the role of negative indexes was changed in the commit 07aef23282374b4a3e630d6ca1968b6f51aa047a, but I forgot to make analogous fix in `isRowAfterRenderedRow` and `isColumnAfterRenderedColumns`. This is my attempt to fix that.

* fix typo in JSDoc

Co-Authored-By: Piotr Laszczkowski <swistach@users.noreply.github.com>