bugfix: The border is misaligned when the table is scrolled
as pointed out in https://github.com/handsontable/handsontable/pull/6157#issuecomment-532590558

The problem was in CSS classes that were added to HOT containers *after* the borders are rendered. I needed to change the border rendering function to be called later. Also, I reinforced the WOT tests by making it test this troublesome behavior.