cell editing Undo Redo doesn't work for column setup with data of type function
Cell Column Data is defined as such:

interface ColumnSettings {
      data: ColumnDataAccessor;
     ...

ColumnDataAccessor is defined as such:
 type ColumnDataAccessor = string | ((row:any, val?:any) => string | void) | number;
   
Undo redo functions deep clones the action ([row, ColumnDataAccessor, prev value, new value]), deepClone is essentially a simple JSON parsing of the action and will fail to convert the second item of the action (ColumnDataAccessor) if it is of type function, the result will be null. This prevent the undo redo machinery to find find the cell to set.


One fix is to simply re-assign the ColumnDataAccessor function from the the source action.

const data = deepClone(this.changes);
if (typeof this.changes[0][1] === "function")
  data[0][1] = this.changes[0][1];

Tested in our code.