Add LazyGridMap structure, holder for meta data
Added LazyGridMap structure designed as holder for column and cell meta
objects. The structure is a refactored version of the structure
currently used in the latest version of Handsontable, where "key"s for
values are volatile. Volatile means that the "key" can change after
altering operations (inserting or removing rows) it is required that
"key" passed as physical row/column has to be glued to the specific
record. Additionally, new npm command was added. `npm run
test:e2e.verbose` runs tests in the CLI but with printing test names.