switch Walkontable tests to use new `walkontable.css`
In `walkontable.test.css` only leave the styles that support the test runner

Motivation for this change is that the stylesheets were separate in the past, which required double maintenance and have lead to inconsistent testing in Walkontable of things that exhibit different behavior in Handsontable. This has been discussed in https://handsoncode.slack.com/archives/C4PV4S33Q/p1571751648008000.

Walkontable tests were changed to adapt to the new styles, because some discrepancies between the old and the new styles have surfaced, as expected. Specifically:

- I have added "handsontable" and "htCore" classes to the elements created in the spec files, to make it equivalent with the classes added by `src/tableView.js`.
- `border.spec.js` was changed because the use of $wrapper and $container was confusingly different than in the other tests
- `preventOverflow.spec.js` was changed because the old Walkontable stylesheet was significantly different than in the actual stylesheet used by Handsontable
- `table.spec.js` was changed, because the `border-left-width` in the old Walkontable stylesheet was different than in the actual stylesheet used by Handsontable