fix: do not render cells on master table that would be always overlapped by fixedRowsTop overlay
because that is bad performance but also affects testing. Before this change, if the table had less rows that fit on screen, it would change RowsCalculator.startRow to be 0. Now startRow will consistently always be equal or larger than fixedRowsTop.

This allows me to fix inconsistent rendering of borders at overlay edges in non-scrollable tables