Move common stylesheets from `handsontable.css` to `walkontable… (#6378)
* remove WOT file `bootstrap.css`

because it was not used anywhere

* rename old `walkontable.css` to `walkontable.test.css`

to emphasize that this file is only supposed to contain styles that support tests.

In the next commits, I will create a new file `walkontable.css` and move there the common parts between Handsontable and Walkontable

* split `handsontable.css` into `handsontable.css` and new `walkontable.css`

because the styles that directly affect Walkontable should be defined closer to Walkontable source and tests

* switch Walkontable tests to use new `walkontable.css`

In `walkontable.test.css` only leave the styles that support the test runner

Motivation for this change is that the stylesheets were separate in the past, which required double maintenance and have lead to inconsistent testing in Walkontable of things that exhibit different behavior in Handsontable. This has been discussed in https://handsoncode.slack.com/archives/C4PV4S33Q/p1571751648008000.

Walkontable tests were changed to adapt to the new styles, because some discrepancies between the old and the new styles have surfaced, as expected. Specifically:

- I have added "handsontable" and "htCore" classes to the elements created in the spec files, to make it equivalent with the classes added by `src/tableView.js`.
- `border.spec.js` was changed because the use of $wrapper and $container was confusingly different than in the other tests
- `preventOverflow.spec.js` was changed because the old Walkontable stylesheet was significantly different than in the actual stylesheet used by Handsontable
- `table.spec.js` was changed, because the `border-left-width` in the old Walkontable stylesheet was different than in the actual stylesheet used by Handsontable

* change classes that are dynamically added to be explicit

because otherwise they are not findable in code. Motivation: https://handsoncode.slack.com/archives/C4PV4S33Q/p1571824566001100

this function was unnecessarily complex (https://en.wikipedia.org/wiki/You_aren%27t_gonna_need_it)

* removed unused CSS classes

* fix the formatting of the selector

because Prettier unexpectedly broke it into multiple lines

* coding style: add empty lines at the bottom of the file

as suggested in the code review https://github.com/handsontable/handsontable/pull/6378#pullrequestreview-307188537

Co-Authored-By: Piotr Laszczkowski <swistach@users.noreply.github.com>