Don't render border lines at the edges with row and column headers
This is according to the "Borders" spec in the KB, which says:

Edges with row and column headers

If the cell is a neighbor of a row of column header, the border of that cell is not rendered at the common edge with the header.

Exception: selection highlights are displayed atop the header edges.