Fix onCellMouseUp and onCellDblClick Walkontable hooks
The Walkontable had a bug where `onCellMouseUp` hook was fired for all
mouse buttons except RMB. It was not consistent with `onCellMouseDown`
hook, which fired always.

The behavior of the `onCellDblClick` and `onCellCornerDblClick` is
changed. From now on, the doubleclick is fired only when it's triggered
by LMB. Previously the hooks are triggered on any button except RMB.

Issue: #6507