Fix navigation by arrow keys for the ContextMenu
I've specified more precisely what input element blocks the key
navigation on the ContextMenu menu instance. Only inputs which are
mounted in the menu's parent element are considered.

Otherwise, there is a conflict with other input elements that are
focused e.q by the CopyPaste plugin, which met that condition.

Issue: #6549