change: inline "adjustElementsPositions" in "redrawClones"
because I think that the key in solving the "Case 1", "Case 3" and "Case 4" problems in https://github.com/handsontable/handsontable/issues/6467#issuecomment-561675300 lies in the fact that adjustElementsPosition was done irrespective of redrewClone.

the next step will be to move "adjustElementsPosition" before "redrawClone"

Note that in this change, I needed to change one test (in "selectAll.spec.js"). I believe that the previous test was actually wrong and the new test fixture is the correct one, because there are less rows drawn.