change: add a new border property "strokeAlignment" to determine how the border of a selection is rendered
previously, the className "current" had a magical hard-coded consequence of changing the border to be rendered inside of the cell.

From now on, this behavior is defined by setting a property `strokeAlignment: 'inside'` on the border settings.

The name `strokeAlignment` is inspired by the new SVG Strokes spec proposal: https://svgwg.org/specs/strokes/#SpecifyingStrokeAlignment