Extracted Cell-Meta logic from Core to separate module (#6254)
* Extracted Cell-Meta logic from Core to separate module

Refactored Cell-Meta. It involves extracting all cell-meta logic from
Core module to separate modules. Currently, the idea has stopped at that
cell-meta module was placed into `dataMap/metaManager` sub-module. The
metaManager in that case will be as close as possible to source data and
record translator logic.

The Handsontable uses 4 layers of cell-meta objects and it was extracted
in the same manner.
 - GlobalMeta - global object meta (prototype object) were changing
   values affects sub-metas;
 - TableMeta - the instance of the GlobalMeta, holder for properties for
   table only. Changing values in that object doesn't affect any other
   layers;
 - ColumnMeta - an object which inherits all properties from GlobalMeta.
   Changing the values of this layer affects cell-meta objects;
 - CellMeta - the instance of ColumnMeta, which creates the cell-meta
   object.

Co-authored-by: Wojciech Czerniak <wojciech.czerniak@gmail.com>
Co-authored-by: Anna Cieślik <scarletfog@gmail.com>