change: remove dynamic addition/removal of right and bottom borders of top and left overlays (#6646)
This commit removes .innerBorderLeft, .innerBorderTop classes, which were added dynamically depending on the scroll position and 
the configuration of the headers. These classes were a source of many needs for workarounds.

One such workaround was in a commit made 3 years ago: c315eaf176ec12f1cecfc43d1c4ae8c7b37ff02b, 
which was unfortunate in the terms of architecture. The external settings object was modified in a class constructor. 
First and foremost, Walkontable should treat that object as read-only.

The original problem #3850, which was addressed by c315eaf176ec12f1cecfc43d1c4ae8c7b37ff02b, is now fixed by 
the removal of .innerBorderLeft, .innerBorderTop.

After this change, the borders are rendered on the overlay, even if there are only header cells and no fixed cells. 
The line is, however, not visible, because the right padding excludes lines that are directly on the header line. 
This is a difference from what is expected in https://github.com/handsontable/handsontable/issues/6651#issuecomment-577085656. 
I will fix that in the consecutive commits.