code style: remove unnecessary property assignment, add a hack for NestedHeaders
As suggested in https://github.com/handsontable/handsontable/pull/6157#discussion_r370575387, the properties are not used.

However, one test in the NestedHeaders plugin fails if I remove reading of the 'columnHeaders' option from here. I can't understand why is that, therefore I am committing this as is. Probably there is some incorrect assumption in `NestedHeaders` that `columnHeaders` callback is called before `NestedHeaders` initialize and that assumption might have been broken in my refactors of Walkontable in PR #6157.