Add test files to babel transpiling process (#6715)
* Remove test files from babel transpiling

"commonjs_e2e" environment is used only for tests so ignore files
don't match here.

Remove "jest-matcher-utils" as utils which formats diff between the two
strings. The module is designed to run on Node env it pulls to much
dependencies to bundled file. It seems that custom normalizer which is
used in browser env is enough for NodeJS.

Issue: #6713