change: drawing of borders on the gridlines (#6646, #6673, #6671, #6651)
Bugfix #6646 (Loss of right border of rowHeader while scrolling when fixedRowsTop and fixedRowsBottom
are enabled). The solution was to remove various factors that caused varied size of headers.
There are no longer classes `innerBorderTop` and `innerBorderLeft`. This change simplified the code,
but could have affected various plugins, such NestedHeaders. All plugins should be tested to be sure
that there are no regressions.

Bugfix #6673. First column is now 1px wider than before (as reported by outerWidth), to accommodate for
the fact that it draws border on the left.

Bugfix #6671 (selection flickers when autofill is on last stretched column). The cause of the problem
was that the corner position did not take into the account the scrollbar width

Feature #6651. The border is rendered on the gridline according to the new specification, and the frozen
line has a new color #5d6365.

Additional changes:

1. in `walkontable.css`, change `line-height` of `.colHeader` to a slightly smaller value that results in
the same row-height in Firefox as in Chrome on Windows

2. in `walkontable.css`, removed declaration of default <col> width 50px. Now, the same value comes from
Walkontable setting `defaultColumnWidth`. The only way to overwrite it is by providing the new value in the JS
settings object. This is more consistent.

3. make AutoColumnSize results deterministic across all tested browsers. In tests, use `toEqual` instead
of `arrayContaining` or `toBeAroundValue`.

4. in HeaderTooltips, hardcode the text width threshold for the `onlyTrimmed` to 50px. This change is
explained in a comment. Related issue: #6708

5. fix rendering of GhostTable. Before this commit, GhostTable was sometimes rendered without borders
on all sides

6. in all places where 1px is added or subtracted to accommodate for the gridline, use a constant
GRIDLINE_WIDTH defined in `utils/gridline.js`

7. in frozen line tests, change `elem.innerText` to `elem.textContent`, because the former was sometimes
empty in Safari

8. performance: `getCellFn` is now provided as configuration to `BorderRenderer`, instead of as
a parameter to `BorderRenderer.render`