Divide border renderer into small modules
This is a proposition (a seed of idea) of the refactoring BorderRenderer,
which, after developing, has grown into massive module with a different
business logic and DOM manipulation.

Here I've created a SvgElement module, which encapsulates (or should
encapsulate) access to raw DOM structure. The ClientRect keeps as close
as possible all properties about sizing and positioning into one place.