Reorganize the code of how the border style classes are managed
The BorderStyle classes were used in the Selection classes. They should
be encapsulated within the "highlight" module. So the code was adjusted
to keep that in mind. The prototype from the BorderStyle class is no
more explicitly accessed while style updating. That form of
implementation was hidden under the "getCommonBorderStyle" method of
the Highlight module.

The method "updateBorderStyleFromSettings" was renamed to a more generic
name ("updateBorderStyle"). The method accepts a specific object which
can be passed not only form settings.

I added "selectionStyle" option to the settings meta schema file. This
will allow generating documentation for that option.

The "updateBorderStyle" function is moved to the "utils" modules.