change: re-add ability to reset custom `selectionStyle` to the defaults
The previous commit removed ability to switch from a custom `selectionStyle` to the defaults, because it only checked for defined (truthy) properties of the provided object. This commit allows resetting to defaults using an empty object. This way was agreed with @budnix on Slack: https://handsoncode.slack.com/archives/D4S007ZD5/p1583400117001100

this commit also fixes some typos in JSDoc