fix: Incomplete border on the overlay edge when the border range contains a merged cell (#6760)
the problem came from the fact that borders of certain selections (current cell, fill) were not transformed by merged cells plugin. Transformation is needed to expand the cell to a range that fully contains any merged cells. Such expanded range must be forwarded to Walkontable selection, because Walkontable is not aware of merged cells/rowspan/colspan.

this solution is implemented using cell meta, because that allows to decouple autofill and mergeCells plugins