Feature/issue 6911 (#7006)
It seems that the Walkontable Table class has a logical bug in the isRowBeforeRenderedRows, isRowAfterRenderedRows, 
isColumnBeforeRenderedColumns and 
isColumnAfterRenderedColumns method in conjunction with 
hidden indexes. The methods among other things are 
responsible for the logic behind getCell method.

I adjusted those methods to support header coordinates. Generally,
the fix is about adding an additional condition for checking negative 
values (header coordinates). The presence of the header existence 
can be checked only in a case when the overlay is rendered from an
index 0 (first renderable index) or if there are no rows or columns
(first renderable index is then equal to -1). When the overlay is 
enabled by fixedRowsTop and counterparts options then the master
overlay is shifted inwards so its first renderable index starts from 2
(depends on fixedRowsTop etc. settings). Hence, the master overlay
doesn't have headers - in reality, they are hidden and offset from
the top-left position (side effect).

Issue: #6911

Co-authored-by: Wojciech Szymański <141330+wszymanski@users.noreply.github.com>