Fix headers size update after dbl-click on resize handler
Fixed the issue by leaving the `beforeColumnResize` and
`beforeRowResize` listeners no matter if the plugin is disabled or so.
This allows resize handler to calculate the new size after the header
double-click. If the AutoRowSize or AutoColumnSize plugin will be
no available the headers will stay unchanged after double-click.

Issue: #6755